refactor, comments, clean up
